backend/requirements.txt
Django
channels
channels-redis
daphne
djangorestframework
djangorestframework-simplejwt 
django-cors-headers
django-oauth-toolkit
requests
python-decouple
pyjwt
cryptography
pyotp
Pillow